Renamed files, removed multiline, added the raw output, added matchlines.
